+++ src/check-file-names.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const npmInstall = require('./npm-install');
 const pkg = require('../package.json');
-const utils = require('./lib/utils');
+const configHelper = require('./lib/helper-config.js');
 
 function buildConfig(parameters, globals) {
   const defaults = require('./check-file-names-defaults');
@@ -14,13 +14,18 @@ function buildConfig(parameters, globals) {
     globals.sourceFiles = defaults.paramCase;
   }
 
-  return utils.mergeParameters(defaults, parameters);
+  return configHelper.mergeParameters(defaults, parameters);
 }
 
 /**
  * @module tasks/checkFileNames
  */
 module.exports = {
+
+  dependencies: _.pick(pkg.devDependencies, [
+    'gulp-check-file-naming-convention',
+    'event-stream'
+  ]),
 
   /**
    * @function
@@ -43,10 +48,7 @@ module.exports = {
   register(gulp, taskName, parameters, globals) {
     const installDependenciesTaskName = taskName + '-install-dependencies';
     npmInstall.register(gulp, installDependenciesTaskName, {
-      dependencies: _.pick(pkg.devDependencies, [
-        'gulp-check-file-naming-convention',
-        'event-stream'
-      ])
+      dependencies: this.dependencies
     });
 
     gulp.task(taskName, [installDependenciesTaskName], () => {