+++ src/cli/generate.js
@@ -3,6 +3,7 @@
 const fs = require('../lib/fs');
 const path = require('path');
 const pkg = require('../../package.json');
+const yargsHelper = require('../lib/helper-yargs');
 
 /**
  * @module cli/generate
@@ -24,14 +25,7 @@ module.exports = {
    * @param {Object} [yargs] Yargs instance to use
    */
   run(args, yargs) {
-    yargs = yargs || require('yargs'); // eslint-disable-line no-param-reassign
-
-    yargs
-      .exitProcess(false)
-      .showHelpOnFail(false)
-      .help('h')
-      .alias('h', 'help')
-      .strict()
+    yargs = yargsHelper.initYargs(yargs) // eslint-disable-line no-param-reassign
       .options({
         n: {
           alias: 'name',
@@ -47,36 +41,26 @@ module.exports = {
       })
       .usage('Usage: urbanjs generate -n clean-project -f');
 
-    let argv;
-    try {
-      argv = yargs.parse(args);
-    } catch (err) {
-      // most of the time an unknown option causes this error
-      // let's show the help how to use urbanjs correctly
-      yargs.showHelp();
+    let force;
+    let projectName;
+    let folderPath;
 
-      return Promise.reject(err);
-    }
+    return yargsHelper.parseArgs(yargs, args)
+      .then(argv => {
+        force = argv.force;
+        projectName = argv.name;
+        folderPath = path.join(process.cwd(), projectName);
 
-    const force = argv.force;
-    let projectName = argv.name;
-    let folderPath = path.join(process.cwd(), projectName);
+        if (path.isAbsolute(projectName)) {
+          folderPath = projectName;
+          projectName = path.basename(projectName);
+        }
 
-    if (path.isAbsolute(projectName)) {
-      folderPath = projectName;
-      projectName = path.basename(projectName);
-    }
-
-    let generationProcess = projectName
-      ? Promise.resolve()
-      : Promise.reject(new Error(`The given name is invalid: ${argv.name}`));
-
-    if (force) {
-      generationProcess = generationProcess
-        .then(() => fs.delete(folderPath));
-    }
-
-    generationProcess = generationProcess
+        if (!projectName) {
+          throw new Error(`The given name is invalid: ${argv.name}`);
+        }
+      })
+      .then(() => force && fs.delete(folderPath))
       .then(() => fs.exists(folderPath))
       .then(exists => {
         if (exists) {
@@ -85,9 +69,7 @@ module.exports = {
             force ? ' and cannot be deleted.' : '. Use -f to force to delete it.'
           ].join(''));
         }
-      });
-
-    generationProcess = generationProcess
+      })
       .then(() => {
         const packageJSON = {
           name: projectName,
@@ -109,13 +91,25 @@ module.exports = {
         packageJSON.devDependencies[pkg.name] = '^' + pkg.version;
 
         return Promise.all([
-          fs.writeFile(path.join(folderPath, 'docs/main.js'), '// write here your custom jsdoc documentation...\n'),
+          fs.writeFile(
+            path.join(folderPath, 'docs/main.js'),
+            '// write here your custom jsdoc documentation...\n'
+          ),
 
-          fs.writeFile(path.join(folderPath, 'package.json'), JSON.stringify(packageJSON, null, '  ')),
+          fs.writeFile(
+            path.join(folderPath, 'package.json'),
+            JSON.stringify(packageJSON, null, '  ')
+          ),
 
-          fs.writeFile(path.join(folderPath, 'src/index.js'), '// let\'s get started...\n'),
+          fs.writeFile(
+            path.join(folderPath, 'src/index.js'),
+            '// let\'s get started...\n'
+          ),
 
-          fs.writeFile(path.join(folderPath, 'README.md'), `# ${projectName}\n`),
+          fs.writeFile(
+            path.join(folderPath, 'README.md'),
+            `# ${projectName}\n`
+          ),
 
           fs.readFile(path.join(__dirname, '../../.editorconfig'))
             .then(content => fs.writeFile(path.join(folderPath, '.editorconfig'), content)),
@@ -133,16 +127,21 @@ module.exports = {
             .then(content => fs.writeFile(path.join(folderPath, 'gulpfile.js'), content)),
 
           fs.readFile(path.join(__dirname, '../../docs/__fixtures__/static/main.css'))
-            .then(content => fs.writeFile(path.join(folderPath, 'docs/__fixtures__/static/main.css'), content)),
+            .then(content => fs.writeFile(
+              path.join(folderPath, 'docs/__fixtures__/static/main.css'),
+              content
+            )),
 
           fs.readFile(path.join(__dirname, '../../docs/__fixtures__/layout.html'))
-            .then(content => fs.writeFile(path.join(folderPath, 'docs/__fixtures__/layout.html'), content))
+            .then(content => fs.writeFile(
+              path.join(folderPath, 'docs/__fixtures__/layout.html'),
+              content
+            ))
         ]);
-      });
-
-    return generationProcess
+      })
       .then(() => {
-        console.log('Project skeleton has been successfully generated.'); // eslint-disable-line no-console
+        console.log(// eslint-disable-line no-console
+          'Project skeleton has been successfully generated.');
       })
       .catch(err => {
         console.error(// eslint-disable-line no-console