+++ src/cli/index.js
@@ -1,7 +1,9 @@
 'use strict';
 
 const generate = require('./generate');
+const installDependencies = require('./install-dependencies');
 const pkg = require('../../package.json');
+const yargsHelper = require('../lib/helper-yargs');
 
 /**
  * @module cli/index
@@ -24,51 +26,30 @@ module.exports = {
    * @param {Object} [yargs] Yargs instance to use
    */
   run(args, yargs) {
-    yargs = yargs || require('yargs'); // eslint-disable-line no-param-reassign
-
-    let commandExecution = null;
-    const runCommand = (y, command) => {
-      commandExecution = command.run(args, y);
-    };
-
-    yargs
-      .exitProcess(false)
-      .showHelpOnFail(false)
-      .help('h')
-      .alias('h', 'help')
+    yargs = yargsHelper.initYargs(yargs) // eslint-disable-line no-param-reassign
       .version(pkg.version)
       .alias('v', 'version')
-      .strict()
-      .usage('Usage: urbanjs <command>')
-      .command(
-        'generate',
-        'Generates a skeleton for your next project',
-        y => runCommand(y, generate)
-      );
+      .usage('Usage: urbanjs <command>');
 
-    let argv;
-    try {
-      argv = yargs.parse(args);
-    } catch (err) {
-      // the command is responsible for showing help if an execution has been started
-      if (!commandExecution) {
-        // most of the time an unknown option causes this error
-        // let's show the help how to use urbanjs correctly
-        yargs.showHelp();
+    let commandExecution = null;
+    [
+      [generate, 'generate', 'Generates a skeleton for your next project'],
+      [installDependencies, 'install-dependencies', 'Installs the dependencies of the given tasks']
+    ].forEach(command => yargs.command(command[1], command[2], y => {
+      commandExecution = command[0].run(args, y);
+    }));
+
+    return yargsHelper.parseArgs(yargs, args).then(argv => {
+      if (argv.v) {
+        // yargs handles this option synchronously
+        return Promise.resolve();
+      } else if (commandExecution) {
+        // return the promise of the command
+        return commandExecution;
       }
 
-      return Promise.reject(err);
-    }
-
-    if (argv.v || argv.h) {
-      // yargs handles these options synchronously
-      return Promise.resolve();
-    } else if (commandExecution) {
-      // return the promise of the command
-      return commandExecution;
-    }
-
-    yargs.showHelp();
-    return Promise.reject(new Error('Invalid argument'));
+      yargs.showHelp();
+      return Promise.reject(new Error('Invalid argument'));
+    });
   }
 };