+++ src/jscs.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 const npmInstall = require('./npm-install');
 const path = require('path');
 const pkg = require('../package.json');
-const utils = require('./lib/utils');
+const configHelper = require('./lib/helper-config.js');
 
 function buildConfig(parameters, globals) {
   const defaults = require('./jscs-defaults');
@@ -15,13 +15,18 @@ function buildConfig(parameters, globals) {
     globals.sourceFiles = defaults.files;
   }
 
-  return utils.mergeParameters(defaults, parameters);
+  return configHelper.mergeParameters(defaults, parameters);
 }
 
 /**
  * @module tasks/jscs
  */
 module.exports = {
+
+  dependencies: _.pick(pkg.devDependencies, [
+    'gulp-jscs',
+    'jscs-jsdoc'
+  ]),
 
   /**
    * @function
@@ -47,10 +52,7 @@ module.exports = {
   register(gulp, taskName, parameters, globals) {
     const installDependenciesTaskName = taskName + '-install-dependencies';
     npmInstall.register(gulp, installDependenciesTaskName, {
-      dependencies: _.pick(pkg.devDependencies, [
-        'gulp-jscs',
-        'jscs-jsdoc'
-      ])
+      dependencies: this.dependencies
     });
 
     const validate = config => {