+++ src/jsdoc.js
@@ -6,7 +6,7 @@ const npmInstall = require('./npm-install');
 const path = require('path');
 const pkg = require('../package.json');
 const shell = require('gulp-shell');
-const utils = require('./lib/utils');
+const configHelper = require('./lib/helper-config.js');
 
 function buildConfig(parameters, globals) {
   const defaults = {
@@ -18,13 +18,22 @@ function buildConfig(parameters, globals) {
     globals.babel = require('./lib/global-babel');
   }
 
-  return utils.mergeParameters(defaults, parameters);
+  return configHelper.mergeParameters(defaults, parameters);
 }
 
 /**
  * @module tasks/jsdoc
  */
 module.exports = {
+
+  dependencies: _.pick(pkg.devDependencies, [
+    'babel-core',
+    'babel-preset-es2015',
+    'babel-preset-react',
+    'babel-preset-stage-0',
+    'jest-cli',
+    'jsdoc'
+  ]),
 
   /**
    * @function
@@ -50,14 +59,7 @@ module.exports = {
   register(gulp, taskName, parameters, globals) {
     const installDependenciesTaskName = taskName + '-install-dependencies';
     npmInstall.register(gulp, installDependenciesTaskName, {
-      dependencies: _.pick(pkg.devDependencies, [
-        'babel-core',
-        'babel-preset-es2015',
-        'babel-preset-react',
-        'babel-preset-stage-0',
-        'jest-cli',
-        'jsdoc'
-      ])
+      dependencies: this.dependencies
     });
 
     const cleanUpTaskName = taskName + '-clean';