+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+- Introduce ```install-dependencies``` command
+- Add verbose parameter to ```npm-install``` task
+
 ## 0.4.2 (2016-02-19)
 - Fix ```jest``` tasks (globals are not working in parallel mode)